third_party/WebKit/LayoutTests/http/tests/devtools/console/console-pins-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that console can pin expressions.


Running: testBeforeAdding
No pins

Running: testAddingExpression
Name: document.activeElement
Value: div#div1

Running: testPinUpdatesDynamically
Focusing the div2 on target page.
Name: document.activeElement
Value: div#div2

Running: testNoSideEffectsWhileEditing
Focusing the first pin's editor.
Setting text to: "window.flag = true".
Name: window.flag = true
Value: (...)
window.flag is now: false

Running: testRemoveSinglePin
Name: document.activeElement
Value: div#div2
Name: "Second pin"
Value: "Second pin"

Removing second pin

Name: document.activeElement
Value: div#div2

Running: testRemoveAllPins
No pins
